Rename chatlist to chat_list
